Reorganize synthetic jobs tests, move to a separate dir
