+++ validator/app/src/compute_horde_validator/validator/tests/test_synthetic_jobs/conftest.py
@@ -1,0 +1,137 @@
+import uuid
+
+import bittensor
+import pytest
+import pytest_asyncio
+from compute_horde.executor_class import DEFAULT_EXECUTOR_CLASS
+from compute_horde.miner_client.base import AbstractTransport
+from compute_horde.mv_protocol import miner_requests
+from django.conf import settings
+
+from compute_horde_validator.validator.miner_client import MinerClient
+from compute_horde_validator.validator.models import (
+    Miner,
+    SyntheticJobBatch,
+)
+from compute_horde_validator.validator.tests.transport import MinerSimulationTransport
+
+
+@pytest.fixture
+def miner_hotkey():
+    return "miner_hotkey"
+
+
+@pytest.fixture
+def validator_hotkey():
+    return "validator_hotkey"
+
+
+@pytest.fixture
+def miner_axon_info(miner_hotkey: str):
+    return bittensor.AxonInfo(
+        version=4,
+        ip="ignore",
+        ip_type=4,
+        port=9999,
+        hotkey=miner_hotkey,
+        coldkey=miner_hotkey,
+    )
+
+
+@pytest_asyncio.fixture
+async def miner(miner_hotkey: str):
+    return await Miner.objects.acreate(hotkey=miner_hotkey)
+
+
+@pytest_asyncio.fixture
+async def batch():
+    return await SyntheticJobBatch.objects.acreate(
+        started_at="2021-09-01 00:00:00",
+        accepting_results_until="2021-09-01 00:00:00",
+    )
+
+
+@pytest.fixture
+def keypair():
+    return settings.BITTENSOR_WALLET().get_hotkey()
+
+
+@pytest_asyncio.fixture
+async def transport(miner_hotkey: str):
+    return MinerSimulationTransport(miner_hotkey)
+
+
+@pytest_asyncio.fixture
+async def miner_client(
+    miner_hotkey: str, validator_hotkey: str, keypair, transport: AbstractTransport
+):
+    return MinerClient(
+        miner_address="ignore",
+        my_hotkey=validator_hotkey,
+        miner_hotkey=miner_hotkey,
+        miner_port=9999,
+        job_uuid=None,
+        batch_id=None,
+        keypair=keypair,
+        transport=transport,
+    )
+
+
+@pytest.fixture
+def job_uuid():
+    return uuid.uuid4()
+
+
+@pytest.fixture
+def manifest_message():
+    return miner_requests.V0ExecutorManifestRequest(
+        manifest=miner_requests.ExecutorManifest(
+            executor_classes=[
+                miner_requests.ExecutorClassManifest(executor_class=DEFAULT_EXECUTOR_CLASS, count=1)
+            ]
+        )
+    ).model_dump_json()
+
+
+@pytest.fixture
+def executor_ready_message(job_uuid: uuid.UUID):
+    return miner_requests.V0ExecutorReadyRequest(job_uuid=str(job_uuid)).model_dump_json()
+
+
+@pytest.fixture
+def accept_job_message(job_uuid: uuid.UUID):
+    return miner_requests.V0AcceptJobRequest(job_uuid=str(job_uuid)).model_dump_json()
+
+
+@pytest.fixture
+def decline_job_message(job_uuid: uuid.UUID):
+    return miner_requests.V0DeclineJobRequest(job_uuid=str(job_uuid)).model_dump_json()
+
+
+@pytest.fixture
+def docker_process_stdout():
+    return "stdout"
+
+
+@pytest.fixture
+def docker_process_stderr():
+    return "stderr"
+
+
+@pytest.fixture
+def job_finish_message(job_uuid: uuid.UUID, docker_process_stdout: str, docker_process_stderr: str):
+    return miner_requests.V0JobFinishedRequest(
+        job_uuid=str(job_uuid),
+        docker_process_stdout=docker_process_stdout,
+        docker_process_stderr=docker_process_stderr,
+    ).model_dump_json()
+
+
+@pytest.fixture
+def job_failed_message(job_uuid: uuid.UUID, docker_process_stdout: str, docker_process_stderr: str):
+    return miner_requests.V0JobFailedRequest(
+        job_uuid=str(job_uuid),
+        docker_process_exit_status=1,
+        docker_process_stdout=docker_process_stdout,
+        docker_process_stderr=docker_process_stderr,
+    ).model_dump_json()
